Aggregation and Duel Mode PBIX pushed
This PBIX contains a use case of loading a large fact table from Snowflake in DirectQuery along with an aggregated table using Import mode. Also, the data model applies Dual model in dimensions. The data source in Snowflake will expire as the Snowflake is in trial mode. Please follow the M query to understand the operations.